+++ src/org/watto/ge/plugin/_Plugin_XXX_Write_DirFile.java
@@ -122,7 +122,7 @@ public class _Plugin_XXX_Write_DirFile extends ArchivePlugin {
 
       // RESETTING GLOBAL VARIABLES
 
-      long arcSize = (int) path.length();
+      long arcSize = path.length();
 
       File sourcePath = getDirectoryFile(path, "idx");
       FileManipulator fm = new FileManipulator(sourcePath, false);
